@@ -19,38 +19,31 @@
 	</organization>
 
 	<properties>
-		<java.version>11</java.version>
-		<maven.compiler.source>11</maven.compiler.source>
-		<maven.compiler.target>11</maven.compiler.target>
+		<java.version>17</java.version>
+		<maven.compiler.source>17</maven.compiler.source>
+		<maven.compiler.target>17</maven.compiler.target>
 		<!-- charset -->
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 		<!-- dependencies -->
-		<owasp.version>6.5.0</owasp.version>
-		<spring.boot.version>2.6.3</spring.boot.version>
-		<spring.cloud.version>2021.0.0</spring.cloud.version>
-		<spring.test.version>5.3.19</spring.test.version>
-		<spring.security.version>5.6.0</spring.security.version>
-		<lombok.version>1.18.22</lombok.version>
-		<springdoc.version>1.6.0</springdoc.version>
-		<mapstruct.version>1.4.2.Final</mapstruct.version>
-		<junit.jupiter.version>5.8.2</junit.jupiter.version>
-		<mockito.version>4.1.0</mockito.version>
+		<owasp.version>8.3.1</owasp.version>
+		<lombok.version>1.18.28</lombok.version>
+		<springdoc.version>2.1.0</springdoc.version>
+		<mapstruct.version>1.5.5.Final</mapstruct.version>
 		<bcpkix.version>1.70</bcpkix.version>
-		<okhttp.version>4.9.1</okhttp.version>
-		<shedlock.version>4.30.0</shedlock.version>
+		<shedlock.version>5.4.0</shedlock.version>
 		<nimbusds.version>9.9.2</nimbusds.version>
-		<dgc.lib.version>1.1.13</dgc.lib.version>
+		<dgc.lib.version>2.0.2</dgc.lib.version>
 		<dgc.bloomfilter.version>0.0.0-29d785a</dgc.bloomfilter.version>
 		<dgc.partialvarriablehashfilter.version>0.0.0-613ab9b</dgc.partialvarriablehashfilter.version>
-		<sap.cloud.sdk.version>3.57.0</sap.cloud.sdk.version>
+		<sap.cloud.sdk.version>4.16.0</sap.cloud.sdk.version>
+		<sap.cloud.logging.version>3.7.0</sap.cloud.logging.version>
 		<jjwt.version>0.11.2</jjwt.version>
 		<!-- plugins -->
-		<plugin.checkstyle.version>3.1.2</plugin.checkstyle.version>
-		<plugin.sonar.version>3.9.0.2155</plugin.sonar.version>
-		<plugin.jacoco.version>0.8.7</plugin.jacoco.version>
+		<plugin.checkstyle.version>3.3.0</plugin.checkstyle.version>
+		<plugin.sonar.version>3.9.1.2184</plugin.sonar.version>
+		<plugin.jacoco.version>0.8.10</plugin.jacoco.version>
 		<plugin.os-maven.version>1.7.0</plugin.os-maven.version>
-		<plugin.surefire.version>3.0.0-M5</plugin.surefire.version>
 		<!-- license -->
 		<license.projectName>EU Digital COVID Certificate Revocation Distribution Service / dgca-revocation-distribution-service
 		</license.projectName>
