@@ -20,8 +20,8 @@
 
 package eu.europa.ec.dgc.revocationdistribution.entity;
 
-import javax.persistence.Entity;
-import javax.persistence.Table;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
