@@ -37,13 +37,13 @@ import eu.europa.ec.dgc.revocationdistribution.repository.PointViewRepository;
 import eu.europa.ec.dgc.revocationdistribution.repository.SliceRepository;
 import eu.europa.ec.dgc.revocationdistribution.repository.VectorViewRepository;
 import eu.europa.ec.dgc.revocationdistribution.utils.HelperFunctions;
+import jakarta.annotation.PostConstruct;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
-import javax.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
